Rebind Caps-Lock to F8 and use it for ace-window
Using <escape> to call ace-window had an odd behaviour when switching
from normal buffers to EXWM ones. The first keystroke after switching
was not being passed to the X client, leading normally to a "buffer is
in read only mode error". Rebinding Caps to a less problematic key
does the trick.