Start Emacs server
Even though I do most of my activities within Emacs and the server is
normally not necessary, I found a case in which having a server
running might be useful. This is when opening files using Firefox. For
this to work, the .desktop file for Emacs which lives in:

/usr/share/applications/emacs.desktop

has to be adapted to call `emacsclient` instead, something like:

  [Desktop Entry]
  Name=Emacs
  GenericName=Text Editor
  Comment=Edit text
  MimeType=text/english;text/plain;text/x-makefile;text/x-c++hdr;text/x-c++src;text/x-chdr;text/x-csrc;text/x-java;text/x-moc;text/x-pascal;text/x-tcl;text/x-tex;application/x-shellscript;text/x-c;text/x-c++;
  Exec=emacsclient %F
  Icon=emacs
  Type=Application
  Terminal=false
  Categories=Development;TextEditor;
  StartupWMClass=Emacs
  Keywords=Text;Editor;