Unbind C-x C-c (save-buffers-kill-emacs)
I rarely quit Emacs as it's my WM and I often accidentally press C-x
C-c when I rapidly realise that the prefix I want to use is C-c and
not C-x. Because of these two, I think that it's not justified to have
a binding for this action. Emacs can be killed using `kill-emacs` or
`save-buffers-kill-emacs`.