Stop using which-key
This package has served well during the ramp-up but now I rarely use
it and most of the time when it pops up it does it because I'm
thinking what to do after executing a prefix, not because I actually
need it.

Also, doing C-h after non-complete key combinations is good enough not
justifying having this package anymore.