Handle always EXWM buffers in lowercase
EXWM buffers get their name from the X window class name. These names
are not consistent across distributions, for instance in Debian
Firefox's is capitalised ("Firefox") but in Arch is not ("firefox"). A
way to be agnostic about this is to always set them to lowercase when
they're assigned to buffer names.