Use the modeline to show ace-window labels
Ace Window cannot draw on EXWM buffers so it makes sense to use
something more consistent (like the modeline) to display the window
labels. Also disable the window blanking (does not work on EXWM
buffers either) and as a complement show that ace-window is active on
the minibuffer as a compensation.