Change icon for Puppet files and modes
The icons provided in the font file-icons [0] don't seem to be
monospaced, so thinner icons may result in misaligned lists for
instance when calling `ivy-switch-buffer'. There are still other icons
like the one used for Help buffers that have the same problems but as
I edit Puppet code very often "fixing" this one will make things
already a lot better.

The icon I'm choosing is of course not the best but it's similar
enough to the Puppet one.

[0] https://github.com/file-icons/icons/blob/master/charmap.md