Always ask if the repo being cloned is the pushDefault
Same when adding a remote.

This makes more sense as default as when cloning repositories that
require forking to contribute to (the vast majority out there) I don't
want them to be the push remote as I don't normally have push perms
there (but the fork -- to be added later).