Attempt to remember the project after cloning
As project.el-wise normally everything starts when I clone a project,
it'd be nice if the project was remembered as soon as it's
cloned. Unfortunately there's no "after clone hook" so I have to use
this hack for the time being. Unfortunately it'll be called much more
time than needed.