Revert "Enable erc-scrolltobottom-mode"
In some cases it conflicts with EXWM:

error in process filter: exwm-input--fake-last-command: ‘recenter’ing a window that does not display current-buffer.
error in process filter: ‘recenter’ing a window that does not display current-buffer.

(current-buffer) is the ERC buffer that recenters via
post-command-hook but, after switching to an EXWM buffer, it's already
not displayed in any window when called.

To be investigated.

This reverts commit 323d1a48f5d9fd3993824f54d371a6788e8ce426.